Collapse ClusterReceptionist.Settings into ReceptionistSettings
